cpp05-08 linux tested, some not working due to subject asking for c++11 things and only works on bugged macOS compiler

+++ cpp06/ex01/include/Serializer.hpp
@@ -17,6 +17,6 @@ public:
 	~Serializer();
 	Serializer &operator=(const Serializer &src);
 
-	static unsigned long serialize(Data* ptr);
-	static Data* deserialize(unsigned long raw);
+	static uintptr_t serialize(Data* ptr);
+	static Data* deserialize(uintptr_t raw);
 };
