+++ cpp06/ex01/src/Serializer.cpp
@@ -15,11 +15,11 @@ Serializer::~Serializer()
 	std::cout << "Serializer Deconstructor called" << std::endl;
 }
 
-unsigned long Serializer::serialize(Data* ptr)
+uintptr_t Serializer::serialize(Data* ptr)
 {
-	return reinterpret_cast<unsigned long>(ptr);
+	return reinterpret_cast<uintptr_t>(ptr);
 }
-Data* Serializer::deserialize(unsigned long raw)
+Data* Serializer::deserialize(uintptr_t raw)
 {
 	return reinterpret_cast<Data*>(raw);
 }
